+++ src/main/java/org/vivecraft/compatibility/VivecraftCompatibility.java
@@ -5,7 +5,7 @@ import org.bukkit.entity.Enderman;
 
 public interface VivecraftCompatibility {
 
-    void editCreeper(Creeper creeper, int radius);
+    void editCreeper(Creeper creeper, double radius);
 
     void editEnderman(Enderman enderman);
 }
